Issue #15: Fix text for license application
The LICENSE file contains a section about how to apply the terms of
AGPL-3.0-only to source files. The notice contained the statement that
version 3 or later applies. That's not the case, because the source code
of Univention products is licensed under AGPL-3.0-only.

This change fixes the mistake. The source files usually state
AGPL-3.0-only.

Because, LICENSE is the central file with the full copy of the license,
the change is necessary.