fix: adjust test_converting_ox_user_to_non_ox_user_updates_cache_correctly test
The current implementation assumes that the user object will still be present
in the non_ox_mapping database after removing the user. This is not the case
for how the connector works in k8s. A followup ticket has been created to
get a decission on how to handle it.

Issue: univention/dev/internal/team-nubus#1369